backend/requirements.txt
pydantic-ai-slim[bedrock]
amazon-transcribe
